Making compatible with PHP 5.3.2
PHP 5.3.2 assumes method names with the same name as the class is a
constructor, so removed those collisions throughout the code.  Also,
for some reason the autoloader isn't called for dynamic class names
in 5.3.2, so adding a temporary fix to index.php where I don't want
to live without dynamic class names (for now).