Adding foreign key functions
Starting to add functions that allow for easy access to linking recrods.
That is to say, any table which has foreign keys in it will allow for
calls to get_<record>(), where <record> is the name of the table which
the foreign key refers to, or get_<record>_list(), where <record> is
the name of a table which has a linking "has" table with the current
record.