Adding error codes.
In order to better handle specific exceptions and improve upon
codes delivered to the user when there is a problem error codes
have been added to the system.

All codes are defined in api/exception/error_codes.inc.php
Codes are defined by the type of exception which throws them.
Anytime a new exception is thrown in the code base, a corresponding
error code must be included in this file.