Creating a menu in a left extruder.
Part of this update required adding get_flap_css_* methods to util.
This is because the flip-text jquery plugin used by the extruder
plugin isn't smart enough to let us add classes dynamically.
This is how the top extruder flap was painted using the jquery ui
theme, so this hack allows us to paint the left extruder flap in
a similar way.