Improving queue, adding province table.
The queue now makes use of several views so that it can be sorted
and restricted based on the participant's last call status, consent
and primary location.

Also, a province table has been added so that each province can be
associated with a particular site for the purpose of deciding which
site calls a participant.