Working on calendars, tweaked operator interface.
Calendars will now only load data which is visible and use AJAX to
dynamically load additional events when the visible day-span changes.
The site calendar is a work in progress but not yet fully functional.

Also changed the operator's interface.  Removed the menu extruder and
made the operator_assignment their home page.