Adding date stamped and form code to contact forms
Contact forms may now be date-stamped.  If this date is recorded during
data entry then a "consent to contact stamped" event will be added to
the participant (if the event exists).  Additionally, a new "grouping"
column has been added to the participant table which is defined by the
code selected during the data entry phase.