Showing all site's assignments
When showing a parented assignment list all assignments are now shown
even if they do not belong to the user's current site (and that user's
role restricts data to their site).  This is in case a participant is
switched to a different site so that users at that site can see the
participant's full assignment details.