Can now edit passed unassigned appointments and callbacks
Passed appointments and callbacks can be edited so long as they have not been assigned.  This is to address the use-case where the helpline receives a call to reschedule an appointment which was missed (or set a callback).  Helpline staff will be able to reschedule by editing missed appointments/callbacks to a new date (they are not to create new appointments/callbacks).