Fixing possible incorrect assigning of appointment in database\appointment::post_process()
If a user were to make an appointment while in an assignment but before calling the user (a situation
which should never happen in practice!) then when starting the call that future appointment was being
assigned to the current assignment.  There is now a check to make sure not to assign future
appointments to an assignment in the database\appointment::post_process() method.